chore: abstracts blocks
